Refactor: Implement 'List', 'Media' and 'Partners' components, remove unused Section component
